+++ src/RestApi/IOptions.ts
@@ -12,6 +12,13 @@ interface IOptions {
 	contentType?: string;
 	auth: IAuthLikeOptions;
 	version: ApiVersions.V1 | ApiVersions.V2;
+	/**
+	 * Define the user custom client versions which is used for header User-Agent.
+	 * E.g.: { "signageOS_CLI": "1.0.3" }
+	 */
+	clientVersions: {
+		[clientName: string]: string;
+	};
 }
 
 export default IOptions;